pyproject.toml
[project]
name = "straight-connect"
version = "0.1.0"
description = "MCP server for communication service connectors"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
config = "src.config:main"
chat = "src.chat:main"
